move payment options from data to entities
